+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - **Changed**
   - `#135`: As dev/test, I want to exclude `test-runner` and `.eslintrc.js` from npm package.
   - `#136`: As scp user, I want to change `skipIfExists` default to `false`.
+  - `#140`: As scp user, I want to standardize the warning to be `'is undefined'` for `undefined`, `null` or `blank`.
+    - and other comestic changes in error message and warning messages.
 - **Fixed**
   - `#129`: bug: `Error: Error: customFilter is not a function [null]`, when `option.customFilter=null`.
 - **Other**